performance improvement
This is a very slight performance improvement.
Before we would check twice if the bank was open, once at the start of the function and again in Self.GetCurrentTab.

Now if the bank is open and we want to open inventory it's only checked once.
For other tabs the bank is still checked twice but it's still a improvement even if minor.
Performance could be improved further by including the whole Gametabs.GetCurrentTab function inside this one but it would make things unreadable and it's not worth it imo.