chore: new tool to draw polygons on walker maps.
New tool to draw polygons on a map that can then be projected into the minimap to figure out if we or something is inside/outside of the polygon. This technique is heavily used by my zeah rc and mlm scripts but for those I had to make the polygons manually, this makes it better.

There's also a new function to check if a point is between 2 points. This can be useful for things like doors for example.