feat(MMDots): the new dot filters are now working
This is still being tested but everything seems to run fine.

Instead of a simple radius to filter mmdots now we have an array of filters that can be added to each mmdot we are looking for.
This can be circles or polygons, can filter the inner part of said shape or the outter part and can have a timer that deletes the filter when the timer runs out (this will be useful to loot items so we don't constantly check them).

The downside of all this is that this has become super hard to read understand. Hopefully I can simplify it someday in the future.