fix: fixed a typo that made loothabdler pick useless stuff.
Mt formula was wrong and I had it looting any item that it's alch value was more than the GEValue. The result was that it was picking up a lot of useless steel stuff.

I also refactored this in a way that will do less API calls to osrsbox via ItemData is they are not needed.