fix: read notes.
This is not really a big deal but I originally intended that both `StatsClient.Disabled` and `WLSettings.Stats` could toggle the stats client  off.
I accidentally made it so `StatsClient.Disabled` was ignored and overriden with whatever `WLSettings.Stats` had set. It's now fixed, and both variables can be used to toggle the stats client off.