fix: read comments
overrides.simba is now included with wasplib as a blank file. This was done for 2 reasons:
- People sometimes thought the message that overrides.simba was missing was an error and they should worry
- I don't like how wasplib has so many files scattered around simba folder, I want to tidy it up.

`StatsPayload.SetUsername()`was added and will now become the prefered way of setting the username instead of `StatsPayload.Username := '...';`
This is because, yesterday I changed wasp-stats so username can be overwritten at will and with that, It's extremely easy for
people to accidentally clear their stats username because all scripts default to `''`.
To fix that, we will be using `StatsPayload.SetUsername()` from now on and people that want to have it permanently set for all scripts
can add something like this to overrides.simba:
```pascal
begin
  StatsPayload.SetUsername('Torwent');
end;
```
Keep in mind that if you add this in overrides.simba, it will take precedence to whatever you set in the scripts.