feat: dumped coordinates. More info in comments
This update is mostly for devs but might affect everyone slightly.
This brings some of my auto-generated coordinates to WaspLib.

This means 2 things:
- Coordinates might need some adjusting (that's why it might affect everyone).
- Coordinates were added in mass.

What this means is that ~90% of the coordinates for the following objects were added:
- Oak, Willow, Teak, Mahogany, Yew, Magic trees
- Bank booths
- Deposit boxes
- Bank chests

Notice regular trees are missing. This is actually on purpose, there's literally hundreds of Object IDs for regular trees and not all of them are actually "choppable" so I couldn't even use the uptext to dump their coordinates. So a complicated way to do it or doing it hundreds of times was required to dump their coordinates and I decided to skip them.

Anyway, this will make making future scripts much easier as you will have 90% of things easily accessible through wasplib.