fix: massive performance improvements
There's 2 methods that I was doing things in a very slow way, mainly due to knowledge back when I made them and that have just been improved.

One of them (`TRSMainScreen.Filter(TPA)`), only affects resizable and is the reason some scripts hanged. The performance gain in this is just unmeasurable... when it caused scripts to hang (usually fishing scripts) it's because it was doing an operation that would take over 15 minutes to complete, it now takes on average 2ms. So no more script hanging from this.

Another thing that was improved was `TBox.Contains(TPA)` which is  "only" 100x faster now and is used a lot in wasp scripts and wasplib itself.

RSNPCs had SpiritAnglers added and were refactored to match RSObjects