fix(data): monsters data loads faster now
The way monster data was held was reworked to several tiny files in a zip instead of the huge, big json file.

This makes loading monster data much faster, however, the biggest bottleneck is actually price checking each drop, which will be looked into in the future since it's what makes the whole thing seem incredibly slow.

For now, in a human scale, things have become a tiny bit faster but not a very dramatic change.