feat(ItemData): ItemData caching implemented
ItemData uses data from 2 sources, wiki prices, which are dynamic and always changing and OSRSBox which is static data that never changes. This static data is now cached, resulting in some dramatic performance improvements where this was used a lot, e.g. monster drops.

The initial run to fetch data of a certain item is the same speed as before, but consequent runs are pretty much instant.

Example loading abyssal demon drops:
- First run: 10951,73 ms
- Second run: 619,578 ms