feat: added Olly's recorder
feats:
Added a heavily modified version of Olly's simba recorder to allow the usage of remote input and also box filters to hide sensitive player data.
This is disabled by default and for now can only be turned on with `WLSettings.SaveVideo := True;` and the recorder will be called and start recording when the first time `RSClient.IsLoggedIn()` is called and it's `True`.
It's set to record the last 15 seconds, I'll consider more in the future.

fixes:
- moved the logic to hide username in screenshots to a different file.
- added more sawmill colors to work better at prifddinnas