fix: minor tweaks to transport.simba
- I find transport.simba extremely verbose. I decided to change every `WriteLn` to be a `TSRLBaseRecord.DebugLn` so it can at least be turned off.
- refactored rsregionutils.simba to include a method to find a point as well instead of what was in transport.simba

Other changes:
Renamed the compiler definitions of everyone's work that wasn't originally mine to reflect it's original developer.
e.g:
  - `WL_SILVERSCREEN_INCLUDED` > `AERO_SILVERSCREEN_INCLUDED`
  - `WL_MUSHTREE_INCLUDED` > `ANDREW_MUSHTREE_INCLUDED``
It's not as readable as it used to be, but I think it's a good way to tribute the original developer of something no matter how much it changes throughout time.