fix(walker): read notes
When I reworked `Walker.GetClosestPoint()` I for Tempoross script I knew the implementation was flawed but I needed something that worked quickly and was better than the previous one.
The way it was working since then was simply count counting the number or nodes (points) in a path and choosing the ones with least points. This was bad because because it doesn't take into account the distance between each node.
The current implementation is how it should have been in the first place.
The distance of each node in each path is measured and summed to get the total distance of each possible path and then we pick one of the shortest ones with some randomization.

This is why WC script was taking very weird routes to oaks and bank in varrock but this improves all TRSWalkerObject walking and anything else that uses `Walker.GetClosestPoint()`