fix: read notes
Most handlers in wasplib are not auto-included by osr.simba anymore. People that want to use this this will have to manually include them from now on.
This is not necessarily a permanent change, more or less a test to see how this does.
Things that depend on each other were changed in a way that let's you import only the top most dependency, e.g:
- if you need the rsalchhandler you can import only that.
- if you need combathandler, it depends on the loothandler which depends on the RSAlchHandler. You can import only the combathandler which will import the loothandler which will import the rsalchhandler.
- if you need skunks teleports, they depend on his teleport utils, you can import only the rsteleports.simba file.
- if you need skunks transport system which depends on his rsteleports which depend on his teleport utils you can import only transport.simba to import all 3.

This might take some get used to but should make wasplib less cluttered and compile faster by not included things that are completely unnecessary for a certain script.
I'm also considering doing the same thing to some of the more niche interfaces like mushtress, baloons, anvil, etc but let's see how this does for now.

BREAKING CHANGES: a lot of scripts will require to be updated for this. If you get a compile error, update your stuff!