fix(TRSWalkerObjects): performance improvements
I spotted a large bottleneck for TRSWalkerObjects with a large amount of coordinates (e.g. trees, rocks when large maps are loaded) and made some improvements to them. There's still further improvements that can be done but objects that were slow due to this should now be ~50% faster at finding