feat(webber): added downscalling option (read notes)
- The world map WaspLib has is too big and causes problems with some of the developer tools.
This will probably have to be changed someday but for now, this gives you the option to use webber with good performance with a tiny bit loss of accuracy and a pretty large loss in visual quality.
the scaling can be toggled with the `SCALE` boolean contant at the top and if used, I would recommend once you are done, turn it off and see how the changes appear while not scaled down.
- added webgraph stuff from Student