feat(poh): finished. read notes
The POH Handler is finished. At least the original intention I had for it which was an accurate way of knowing a player position on anyone's POH no matter their layout.
Actual walking will be done at a later time, but most people could add it in now by looking at how walker does it.
It was also not my original intention for the first release to have it, but it does include some very basic interaction with some very few "Room Objects".
This only includes the fairy ring, the jewellery box and the pool. I only tested this on a max house so the `Finders` might need adjustments. Also, while only this 3 objects are included as a proof of concept, it's not very hard to add other stuff, like portals, nexus, etc. Do keep in mind though, the syntax for this is subject to change in the future.
Things that changed since the last updates:
- All map and "cached" rooms things were moved into their own `TPOHMap` record.
This is a helper record and you likely will never have to interact with it directly nor none of it's methods.
- A lot of methods were renamed to make more sense
- The file was renamed from pohhandler.simba to poh.simba
- The whole file has been documented thoroughly and includes usage examples of the several things the POH Handler can do.