feat: docs - read notes
With this release I've completely finished my goals for the documentation tools.
I've also fixed some more parsing errors that were around the docs.
The source code comment parser was completely rewritten because:
- I didn't like the way it just dumped all links on a single page. That is okay for a handful of files, but  not anymore.
- With the current size e of SRL-T and WaspLib I think having things into their own sections (which are the same as their directories) makes it easier to navigate.
This thing was and still is written in python. I've read some python in the past but never actually used it for anything until I started modifying the docs tool.
This means:
- The doc tool is probably poorly written. I know for a fact that it's doing more loops than it should have to but it was just easier for me this way and it works which is what matters most.
- There could be bugs. I tested it in both SRL-T and WaspLib documentation and it seems to work fine but I'm not sure there aren't some weird cases.

With that said, the way it works now is:
- each file that has doc comments will have it's own html page created.
- each root directory that has doc pages under it will have a index file created for those pages, let's call them main folder indices.
If a file with documentation matches a main folder index page, it's documentation will be added at the top of that page.
  - For example, optional.simba currently have docs, it's comments are added to the optional page, and at the bottom of the page you have your index to each of the optional files that are documented
- a final index page is created indexing each of the main folder indices.
The final result is pretty neat and organized. All tools under the tools sections, all utils under the utils section, all osr files under the osr section and all otpional docs under the otpional section.