fix: removed the "compability layer"
This was okay when Simba1500 was gonna be a thing, but as soon as it was not, this became a terrible thing that caused a lot of redundancy and had a bunch of methods that were called differently but did the same.
I removed everything that was not in use from it (compiled almost every script to test)
and moved most things that remain to SRL-T, and very few things to other places in WaspLib.

BREAKING CHANGE: This will require the latest SRL-T and some of my scripts will need to be updated.