feat(POH): walking
There's a lot in this commit so I'll break it down:
- Fixes:
  - TRSWalker.Click() had a possible infinite loop which was fixed.
- Refactors:
  - TRSWalker.InternalSetup() now sets up walker randomness. I swear this was already done by default in the past but apparently not.
  - TRSPOHMap was moved into it's own file, very similar to the way walker has it. The POH file grew a lot and since this doesn't depend on anything, having it separate is just easier to work on them.
  - TRSWalkerObjects debugging methods (e.g. `Debug(TRSObject)`) were reworked to not draw the "floorTiles" and "roofTiles". This was pretty needless since it gave no extra info than what we already had with the "cuboids" and it was a big resource hog. In any case, it doesn't matter much, it was only used for debugging.
  - Because of the above, the `TRSWalkerObject.Find()` methods could be cleaned up, still no real change, things are just cleaner.
- Features:
  - TRoomObjects have now debugging methods (not tested)
  - Added colors and rough coordinates to several TRoomObjects that were missing (nexus, exit portal, magic altar, larder, ... None were tested, their offsets probably need tweaking, colors should be okay)
  - TRSPOHHandler can now walk (untested).

Comments:
Adding walking on the POH is one of the few things I've done in a long time that I really think is not well done.
This was an exact copy of what TRSWalker has and I don't like that.
I think that going very far to refactor things so code is not reused in 3 or 4 methods is crazy, but this were 9 methods,
probably more in the future and almost 500 lines of the almost exact same thing with very minimal change.

This brings me to my next point:
Walker should be refactored to have it's positioning separate from it's walking.

This is because I can see this being needed again in the future. Some things that come to mind is if someone wants to make a script for The Gauntlet, Chambers of Xeric or anything else that has a randomly generated map.

However... Reworking TRSWalker to have it's positioning and walking separate would be the BREAKING CHANGE of all BREAKING CHANGES.
It would literally break everything. Everything for me, everything for everyone and it would just not be worth it. It would be much worse than when walker 1.5 was migrated to simba 1.4.

 Therefore, this things are going to remain as is, but when a move to Simba 2000 is done, this should be kept in mind.