feat(tools): s2scompiler
This was a side product of my "wasp-launcher" project. It was growing too large to be easy to maintain and I needed to split it into several files to keep working on it.
At the same time, I need the end result to be a single script.
So I've made this tool which is basically a "Simba to Simba compiler" that compiles multiple scripts into a single one.
I thought more people might find it useful.
No instructions given, should be easy to use, figure it out.
It's also pretty basic and simple, won't understand compiler if directives, keep that in mind.