fix(WaspConfig): improved behaviour for non wasplib scripts
Also added a new simple tool "file2string.simba" that was born out of the wasp-launcher.
It reads an entire simba file and print out a string function that will return the entire file contents.
Might sound a bit weird but this is extremely useful if you want to write simba files with simba itself.
1. Write the file
2. Run the tool on the file;
3. Delete the file
4. Use the returned function from 2. to rewrite the file for someone that doesn't have it