feat(housemap): add the new housemap
Basically a full rewrite of `TPOHMap` renamed to `THouseMap` because I like it more and keeps backwards compatibility.

This is still not usable on it's own for anything useful, will require a small rewrite of the `TPOHHandler`, which will be named `THouseHandler` with the global variable of `House`.
Even if it's not usable yet to walk around the house, you may use it to map your house in advance if you want.