feat(THouseHandler): read notes
THouseHandler should be fully functional now with some exceptions:
- portals in portal chambers are not working at all. There's still some work that has to be done for this, it's not much but it's not done yet.
- a lot of objects information (coordinates, uptext, colors) are still missing.

This last one is not something I'll fully complete on my own probably. I'll be adding the information for the objects that I need as I need them. The hard work which was building this is already done, I'll leave filling this information to the community.
Most of it is more or less self-explanatory but you can find more info on this by searching houseutils.simba for "todo" and read the instructions there if any.

While portals still need some extra work to work, you may submit colors for them if you feel like it.

Lastly, there's a House handler form example in the templates folder now