feat: Update scriptform.simba to auto-add / auto-save / track last used account
Big change to scriptform.simba. Many users start scripts without clicking add account or update account and scripts will not login or relog because of it.

This change will recognize when any of the 4 inputs where a user types has a change and will dynamically write those changes to the credentials.simba file. If a username is new/unique, a new account is automatically added and the selected account in the current account dropdown is automatically updated. If the username stays the same, but other details such as worlds has been updated, only the changed information is updated for the already existing account.

Metadata in the form of "// lastUsed=prosimbauser" is written to the credentials file and is also dynamically updated so that simba will remember the last account used and will load that account information on the next script start.

While not necessary, the Add Account and Update Account buttons still exist and their functionality is preserved for those with the muscle memory to use them.

Tested on several scripts including those with per-account settings without issue.